move IconButton component into its own file
