+++ frontend/src/components/Icon.test.tsx
@@ -1,8 +1,7 @@
-import { describe, expect, test, vi } from "vitest";
-import { fireEvent } from "@testing-library/preact";
+import { describe, expect, test } from "vitest";
 
 import { renderWithProviders } from "../test/renderWithProviders";
-import { Icon, IconButton } from "./Icon";
+import { Icon } from "./Icon";
 
 describe("Icon", () => {
   test("should display an icon", () => {
@@ -11,26 +10,5 @@ describe("Icon", () => {
     );
     const span = getByRole('img') as HTMLElement;
     expect(span.className.trim()).toEqual('bi icon bi-snow3');
-  });
-});
-
-describe('IconButton', () => {
-  test('renders an enabled icon button', () => {
-    const onClick = vi.fn();
-    const { getBySelector } = renderWithProviders(
-      <IconButton name="snow3" onClick={onClick} />
-    );
-    fireEvent.click(getBySelector('a'));
-    expect(onClick).toHaveBeenCalled();
-  });
-
-  test('renders a disabled icon button', () => {
-    const onClick = vi.fn();
-    const { getBySelector } = renderWithProviders(
-      <IconButton name="snow3" onClick={onClick} disabled />
-    );
-    expect(getBySelector('a').classList.contains('disabled')).toEqual(true);
-    fireEvent.click(getBySelector('a'));
-    expect(onClick).not.toHaveBeenCalled();
   });
 });
